Started work on guts of calculator
The JS is moving along.  So far, it can accept numbers and clear the output.  It
also only accepts a single "." character.  My next goal is to keep there from
being multiple operations in a row.  Then begins the herculean task of writing
a recursive-decent parser in Javascript to actually parse out the equation.

I also changed the layout of the form a little.