Almost works like a calculator
I thought I might need to implement a RDS for this and I still might but for
now, I'm getting by without one.  The calculator works for a single calculation
(unless that's a divide first).  I think I may need to look into a stack-based
calculator for this project.  That's a thing, right?